+++ quilt/src/main/java/de/maxhenkel/voicechat/integration/ModMenu.java
@@ -3,6 +3,8 @@ package de.maxhenkel.voicechat.integration;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import de.maxhenkel.voicechat.gui.VoiceChatSettingsScreen;
+import de.maxhenkel.voicechat.integration.clothconfig.ClothConfig;
+import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
 
 public class ModMenu implements ModMenuApi {
 
@@ -16,7 +18,7 @@ public class ModMenu implements ModMenuApi {
     }
 
     private ConfigScreenFactory<?> getClothConfigFactory() {
-        return ClothConfigWrapper::createConfigScreen;
+        return screen -> ClientCompatibilityManager.INSTANCE.getClothConfigIntegration().createConfigScreen(screen);
     }
 
 }